Use debug build; don't relink symlink that is already there

+++ build.sh
@@ -4,5 +4,7 @@ swig -I/home/andrew/Documents/fieldml.api/core/src -package fieldml.jni -outdir 
 javac ./fieldml/jni/*.java
 jar -cf fieldml.jar ./fieldml/jni/*.class
 cd ..
-g++ -shared -o libfieldml.so ./core/src/*.cpp -I/usr/lib/jvm/java-6-openjdk-amd64/include/ -Icore/src/ `pkg-config --cflags libxml-2.0` `pkg-config --libs libxml-2.0` jni/*.cxx -fPIC
-ln -s `pwd`/libfieldml.so libfieldml_jni_0.3.so
+g++ -ggdb -O0 -shared -o libfieldml.so ./core/src/*.cpp -I/usr/lib/jvm/java-6-openjdk-amd64/include/ -Icore/src/ `pkg-config --cflags libxml-2.0` `pkg-config --libs libxml-2.0` jni/*.cxx -fPIC
+if [[ ! -e libfieldml_jni_0.3.so ]]; then
+  ln -s `pwd`/libfieldml.so libfieldml_jni_0.3.so
+fi
